# api/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "API",
      "description": "Totoro API Information",
      "version": "1.0"
    },
    "securityDefinitions": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "app_id"
      },
      "bearerAuth": {
        "name": "Authorization",
        "in": "header",
        "type": "apiKey",
        "description": "JWT Authorization header"
      }
    },
    "security": [{ "ApiKeyAuth": [] }, { "bearerAuth": [] }],
...
    "paths": {
       "/api/users": {
           "get": {
           "tags": ["Users"],
           "x-swagger-router-controller": "users",
           "operationId": "getAllUsers",
           "parameters": [
              {
                 "name": "status",
                 "in": "query",
                 "type": "string",
                 "required": false,
                 "enum": ["actived", "disabled", "deleted"]
              }
              ],
              "responses":{
                 "default":{
                    "description":"Unexpected error",
                    "schema":{
                       "$ref":"#/definitions/Error"
                    }
                 },
                 "200":{
                    "description":"Successful request.",
                    "schema":{
                       "$ref":"#/definitions/getAllUsers"
                    }
                 }
              }
           }
        },
...
        "/api/admins": {
           "get": {
           "tags": ["Admins"],
           "x-swagger-router-controller": "admins",
           "operationId": "getAllAdmins",
           "parameters": [
              {
                 "name": "status",
                 "in": "query",
                 "type": "string",
                 "required": false,
                 "enum": ["actived", "disabled", "deleted"]
              },
              {
                 "name": "role",
                 "in": "query",
                 "type": "string",
                 "required": false,
                 "enum": ["Moderateur", "Administrateur", "Comptable"]
              }
              ],
              "responses":{
                 "default":{
                    "description":"Unexpected error",
                    "schema":{
                       "$ref":"#/definitions/Error"
                    }
                 },
                 "200":{
                    "description":"Successful request.",
                    "schema":{
                       "$ref":"#/definitions/getAllAdmins"
                    }
                 }
              }
           },
           "post": {
              "tags": ["Admins"],
              "x-swagger-router-controller": "admins",
              "operationId": "postAdmin",
              "parameters": [
                {
                  "name": "data",
                  "in": "body",
                  "description": "Create Admin",
                  "required": true,
                  "schema": {
                    "$ref": "#/definitions/postAdmin"
                  }
                }
              ],
              "responses": {
                "default": {
                  "description": "Unexpected error",
                  "schema": {
                    "$ref": "#/definitions/Error"
                  }
                },
                "201": {
                  "description": "Admin Created"
                }
              }
           }
        },
...
        "/api/litigations": {
           "get": {
           "tags": ["Litigations"],
           "x-swagger-router-controller": "litigations",
           "operationId": "getAllLitigations",
           "parameters": [
              {
                 "name": "status",
                 "in": "query",
                 "type": "string",
                 "required": false,
                 "enum": ["closed", "deleted"]
              }
              ],
              "responses":{
                 "default":{
                    "description":"Unexpected error",
                    "schema":{
                       "$ref":"#/definitions/Error"
                    }
                 },
                 "200":{
                    "description":"Successful request.",
                    "schema":{
                       "$ref":"#/definitions/getAllLitigations"
                    }
                 }
              }
           },
           "post": {
              "tags": ["Litigations"],
              "x-swagger-router-controller": "litigations",
              "operationId": "postLitigation",
              "parameters": [
                {
                  "name": "data",
                  "in": "body",
                  "description": "Create Litigation",
                  "required": true,
                  "schema": {
                    "$ref": "#/definitions/postLitigation"
                  }
                }
              ],
              "responses": {
                "default": {
                  "description": "Unexpected error",
                  "schema": {
                    "$ref": "#/definitions/Error"
                  }
                },
                "201": {
                  "description": "Litigation Created"
                }
              }
           }
        },
...
        "/api/applications": {
           "get": {
           "tags": ["Applications"],
           "x-swagger-router-controller": "applications",
           "operationId": "getAllApplications",
           "parameters": [
             {
                "name": "status",
                "in": "query",
                "type": "string",
                "required": false,
                "enum": ["actived", "deleted"]
             }
             ],
              "responses":{
                 "default":{
                    "description":"Unexpected error",
                    "schema":{
                       "$ref":"#/definitions/Error"
                    }
                 },
                 "200":{
                    "description":"Successful request.",
                    "schema":{
                       "$ref":"#/definitions/getAllApplications"
                    }
                 }
              }
           },
           "post": {
              "tags": ["Applications"],
              "x-swagger-router-controller": "applications",
              "operationId": "postApplication",
              "parameters": [
                {
                  "name": "data",
                  "in": "body",
                  "description": "Create Application",
                  "required": true,
                  "schema": {
                    "$ref": "#/definitions/postApplication"
                  }
                }
              ],
              "responses": {
                "default": {
                  "description": "Unexpected error",
                  "schema": {
                    "$ref": "#/definitions/Error"
                  }
                },
                "201": {
                  "description": "Application Created",
                  "schema": {
                    "$ref": "#/definitions/Success"
                  }
                }
              }
           }
        },
...
     },
    "definitions": {
        "Error": {
           "required": ["message"],
           "properties": {
              "entity": {
                 "type": "string"
              },
              "message": {
                 "type": "string"
              }
           }
        },
        "Success": {
           "required": ["message"],
           "properties": {
              "entity": {
                 "type": "string"
              },
              "message": {
                 "type": "string"
              }
           }
        },
        "getAllUsers": {
           "type": "object",
           "properties": {
              "total_rows": { "type": "integer" },
              "data": {
                 "type": "array",
                 "items": {
                    "type": "object",
                    "properties": {
                       "id": { "type": "string" },
                       "username": { "type": "string" },
                       "firstname": { "type": "string" },
                       "lastname": { "type": "string" },
                       "email": { "type": "string" },
                       "birthday": { "type": "string" },
                       "longitude": { "type": "integer" },
                       "latitude": { "type": "integer" },
                       "avatar": { "type": "string" },
                       "bio": { "type": "string" },
                       "phone": { "type": "string" },
                       "total_token": { "type": "integer" },
                       "createdAt": { "type": "string" },
                       "updatedAt": { "type": "string" },
                       "status": {
                          "type": "object",
                          "properties": {
                             "label": { "type": "string" }
                          }
                       },
                       "memberships": {
                          "type": "array",
                          "items": {
                             "type": "object",
                             "properties": {
                                "createdAt": { "type": "string" },
                                "updatedAt": { "type": "string" },
                                "organization": {
                                   "type": "object",
                                   "properties": {
                                      "id": { "type": "string" },
                                      "siren": { "type": "string" },
                                      "siret": { "type": "string" },
                                      "name": { "type": "string" },
                                      "email": { "type": "string" },
                                      "longitude": { "type": "integer" },
                                      "latitude": { "type": "integer" },
                                      "creation_date": { "type": "string" },
                                      "activity": { "type": "string" },
                                      "address": { "type": "string" },
                                      "cp": { "type": "integer" },
                                      "commune": { "type": "string" },
                                      "logo": { "type": "string" },
                                      "description": { "type": "string" },
                                      "link": { "type": "string" },
                                      "phone": { "type": "string" },
                                      "createdAt": { "type": "string" },
                                      "updatedAt": { "type": "string" },
                                      "status": {
                                         "type": "object",
                                         "properties": {
                                            "label": { "type": "string" }
                                         }
                                      }
                                   }
                                },
                                "role": {
                                   "type": "object",
                                   "properties": {
                                      "label": { "type": "string" }
                                   }
                                },
                                "status": {
                                   "type": "object",
                                   "properties": {
                                      "label": { "type": "string" }
                                   }
                                }
                             }
                          }
                       }
                    }
                 }
              }
           }
        },
...
        "getAllAdmins": {
           "type": "object",
           "properties": {
              "total_rows": { "type": "integer" },
              "data": {
                 "type": "array",
                 "items": {
                    "type": "object",
                    "properties": {
                       "id": { "type": "string" },
                       "username": { "type": "string" },
                       "firstname": { "type": "string" },
                       "lastname": { "type": "string" },
                       "email": { "type": "string" },
                       "createdAt": { "type": "string" },
                       "updatedAt": { "type": "string" },
                       "status": {
                          "type": "object",
                          "properties": {
                             "label": { "type": "string" }
                          }
                       },
                       "role": {
                          "type": "object",
                          "properties": {
                             "label": { "type": "string" }
                          }
                       }
                    }
                 }
              }
           }
        },
...
        "postAdmin": {
           "type": "object",
           "properties": {
             "firstName": { "type": "string" },
             "lastName": { "type": "string" },
             "username": { "type": "string" },
             "email": { "type": "string" },
             "role_id": { "type": "string" }
           },
           "required": ["firstName", "lastName", "username", "email", "role_id"],
           "example": {
             "username": "uber",
             "firstname": "Quentin",
             "lastname": "Huber",
             "email": "h.quentin@totoro.fr",
             "role_id": "1dc4364b-d3d1-4db4-a324-a1271e59dfb1"
           }
        },
...
        "postLitigation":{
           "type":"object",
           "properties":{
              "litigation_object_id":{"type":"string"},
              "group_id":{"type":"string"},
              "type":{"type":"boolean"},
              "message":{"type":"string"}
           },
           "required": ["litigation_object_id","group_id","type"],
           "example":{
              "litigation_object_id": "561a448e-5ad9-4dd5-8b08-7cc0419824d1",
              "group_id": "561a448e-5ad9-4dd5-1208-7cc0419824d5",
              "type": 0,
              "message": "Lorem ipsum, bla bla bla"
           }
        },
...
        "getAllLitigations": {
           "type": "object",
           "properties": {
              "total_rows": { "type": "integer" },
              "data": {
                 "type": "array",
                 "items": {
                    "type": "object",
                    "properties": {
                       "id": { "type": "string" },
                       "type": { "type": "boolean" },
                       "message": { "type": "string" },
                       "createdAt": { "type": "string" },
                       "updatedAt": { "type": "string" },
                       "status": {
                          "type": "object",
                          "properties": {
                             "label": { "type": "string" }
                          }
                       },
                       "litigation_object":{
                          "type": "object",
                          "properties": {
                             "label": { "type": "string" },
                             "description": { "type": "string" }
                          }
                       },
                       "mission":{
                          "type": "object",
                          "properties": {
                             "id": { "type": "string" },
                             "createdAt": { "type": "string" },
                             "updatedAt": { "type": "string" },
                             "job": {
                                "type": "object",
                                "properties": {
                                   "id": { "type": "string" },
                                   "title": { "type": "string" },
                                   "description": { "type": "string" },
                                   "participants_max": { "type": "integer" },
                                   "address": { "type": "string" },
                                   "cp": { "type": "integer" },
                                   "commune": { "type": "string" },
                                   "start_date": { "type": "string" },
                                   "end_date": { "type": "string" },
                                   "longitude": { "type": "integer" },
                                   "latitude": { "type": "integer" },
                                   "createdAt": { "type": "string" },
                                   "updatedAt": { "type": "string" },
                                   "author":{
                                      "type": "object",
                                      "properties": {
                                         "createdAt": { "type": "string" },
                                         "updatedAt": { "type": "string" },
                                         "organization": {
                                            "type": "object",
                                            "properties": {
                                               "id": { "type": "string" },
                                               "siren": { "type": "string" },
                                               "siret": { "type": "string" },
                                               "name": { "type": "string" },
                                               "email": { "type": "string" },
                                               "longitude": { "type": "integer" },
                                               "latitude": { "type": "integer" },
                                               "creation_date": { "type": "string" },
                                               "activity": { "type": "string" },
                                               "address": { "type": "string" },
                                               "cp": { "type": "integer" },
                                               "commune": { "type": "string" },
                                               "logo": { "type": "string" },
                                               "description": { "type": "string" },
                                               "link": { "type": "string" },
                                               "phone": { "type": "string" },
                                               "createdAt": { "type": "string" },
                                               "updatedAt": { "type": "string" },
                                               "status": {
                                                  "type": "object",
                                                  "properties": {
                                                     "label": { "type": "string" }
                                                  }
                                               }
                                            }
                                         }
                                      }
                                   },
                                   "difficulty": {
                                      "type": "object",
                                      "properties": {
                                         "level": { "type": "integer" },
                                         "token": { "type": "integer" }
                                      }
                                   },
                                   "participant":{
                                      "type": "object",
                                      "properties": {
                                         "id": { "type": "string" },
                                         "username": { "type": "string" },
                                         "firstname": { "type": "string" },
                                         "lastname": { "type": "string" },
                                         "email": { "type": "string" },
                                         "birthday": { "type": "string" },
                                         "longitude": { "type": "integer" },
                                         "latitude": { "type": "integer" },
                                         "avatar": { "type": "string" },
                                         "bio": { "type": "string" },
                                         "phone": { "type": "string" },
                                         "total_token": { "type": "integer" },
                                         "createdAt": { "type": "string" },
                                         "updatedAt": { "type": "string" },
                                         "status": {
                                            "type": "object",
                                            "properties": {
                                               "label": { "type": "string" }
                                            }
                                         }
                                      } 
                                   },
                                   "status": {
                                      "type": "object",
                                      "properties": {
                                         "label": { "type": "string" }
                                      }
                                   }
                                }
                             }
                          }
                       }
                    }
                 }
              }
           }
        },
...
        "getAllApplications": {
           "type": "object",
           "properties": {
              "total_rows": { "type": "integer" },
              "data": {
                 "type": "array",
                 "items": {
                    "type": "object",
                    "properties": {
                       "id": { "type": "string" },
                       "name": { "type": "string" },
                       "createdAt": { "type": "string" },
                       "updatedAt": { "type": "string" },
                       "status": {
                         "type": "object",
                         "properties": {
                            "label": { "type": "string" }
                         }
                      }
                    }
                 }
              }
           }
        },
...
        "postApplication": {
           "type": "object",
           "properties": {
             "name": { "type": "string" }
           },
           "required": ["name"],
           "example": {
             "name": "Test"
           }
        },
...
     }
  }